Kotlin app tests and some refactoring
