+++ example/android/kotlinlib/1-hello-kotlin/app/src/main/java/com/helloworld/SampleLogic.kt
@@ -1,0 +1,5 @@
+package com.helloworld
+
+object SampleLogic {
+    fun textSize(): Float = 32f
+}
